Updates for python 3 compatibility
The most challenging part of this was, not unexpectedly, the bytes/str
difference; I don't feel sure I've gotten every instance correctly. But,
in my testing this does work for everything. After this commit, the
webserver runs without change (as far as I have been able to tell) on
both python 2.7 and python 3, as well as Tornado version 3.2.1, 5.1, and
6.0.3.

The one thing I know I haven't carefully tested is ttyrec writing.