Webtiles: minimal type annotations to pass static type checking
This is done in a way for maximal python 2.7 compatibility, using
comment annotations and very careful importing; not all versions of py27
will have `typing` even available without installing a package. (Though
the most recent ones should.) However, to actually do the type checking,
mypy should be run on the webserver directory in python 3.7 or so with
Tornado 6, which has full type annotations.

The fact that this passes is a good sign, but this commit only deals
with a bunch of global variables that mypy complained about; everything
else is implicit or inherited from Tornado.