Add requirements.txt tools, formatting, and linting
This commit adds a simplified build setup for the webserver project.
User and developer dependencies are specified in requirements.in/ and
compiled to version-pinned dependency lists in requirements/.

This commit also adds some basic linting and formatting tools. All
formatting warnings are disabled, and should be enabled one by one as
the code is organized. I've also thrown in pytest and coverage, although
I haven't set them up yet.